Separate live update logic from component
This should be initialised at the client side after hydrating the wrapper component
